test: exercise fetch defaults through CLI
